+++ packages/cloud-toolkit/src/make-commands.ts
@@ -1,24 +1,64 @@
 import { Command } from 'commander';
-import os from 'os';
 
 import actionRunner from './dialogs/action-runner';
 import askToOverwriteCustomizations from './dialogs/ask-to-overwrite-customizations';
 import { BusinessError } from './errors';
 import bootstrap from './services/bootstrap';
 import { validateEnvironmentVariables, validateServerUrl } from './services/environment-variables';
+import EventSubscriber from './services/event-subscriber';
+import HttpServer from './services/http-server';
 import packageCustomizations from './services/packager';
-import PathManager from './services/path-manager';
 import publish from './services/publish';
 import { updateTypingsWithCustomizations } from './services/update-typings';
-import { MakeCommands } from './types';
+import {
+  BuildEventSubscriber,
+  BuildHttpServer,
+  EnvironmentVariables,
+  Login,
+  MakeCommands,
+  Spinner,
+} from './types';
+
+export const getOrRefreshEnvironmentVariables = async (
+  login: Login,
+  spinner: Spinner,
+  getEnvironmentVariables: () => Promise<EnvironmentVariables>,
+): Promise<EnvironmentVariables> => {
+  let vars = await getEnvironmentVariables();
+
+  if (!vars.FOREST_AUTH_TOKEN) {
+    await login(spinner);
+    vars = await getEnvironmentVariables();
+  }
+
+  return vars;
+};
+
+const validateAndBuildHttpServer = (
+  envs: EnvironmentVariables,
+  buildHttpServer: BuildHttpServer,
+): HttpServer => {
+  validateEnvironmentVariables(envs);
+
+  return buildHttpServer(envs);
+};
+
+const validateAndBuildEventSubscriber = (
+  envs: EnvironmentVariables,
+  buildEventSubscriber: BuildEventSubscriber,
+): EventSubscriber => {
+  validateEnvironmentVariables(envs);
+
+  return buildEventSubscriber(envs);
+};
 
 export default function makeCommands({
-  getOrRefreshEnvironmentVariables,
-  getEnvironmentVariables,
-  buildHttpServer,
+  bootstrapPathManager,
   buildEventSubscriber,
+  buildHttpServer,
+  spinner,
+  getEnvironmentVariables,
   login,
-  buildSpinner,
 }: MakeCommands): Command {
   // it's very important to use a new instance of Command each time for testing purposes
   const program = new Command();
@@ -29,13 +69,22 @@ export default function makeCommands({
         '(whenever its schema changes)',
     )
     .action(
-      actionRunner(buildSpinner, async spinner => {
+      actionRunner(spinner, async () => {
         spinner.start('Updating typings');
-        const vars = await getOrRefreshEnvironmentVariables();
+        const vars = await getOrRefreshEnvironmentVariables(
+          login,
+          spinner,
+          getEnvironmentVariables,
+        );
         validateEnvironmentVariables(vars);
-        const introspection = await buildHttpServer(vars).getIntrospection();
-        const paths = new PathManager(os.tmpdir(), os.homedir());
-        await updateTypingsWithCustomizations(paths.typingsAfterBootstrapped, introspection);
+        const introspection = await validateAndBuildHttpServer(
+          vars,
+          buildHttpServer,
+        ).getIntrospection();
+        await updateTypingsWithCustomizations(
+          bootstrapPathManager.typingsAfterBootstrapped,
+          introspection,
+        );
         spinner.succeed('Your typings have been updated');
       }),
     );
@@ -48,9 +97,13 @@ export default function makeCommands({
       'Environment secret, you can find it in your environment settings',
     )
     .action(
-      actionRunner(buildSpinner, async (spinner, options: { envSecret: string }) => {
+      actionRunner(spinner, async (options: { envSecret: string }) => {
         spinner.start('Bootstrapping project');
-        const vars = await getOrRefreshEnvironmentVariables();
+        const vars = await getOrRefreshEnvironmentVariables(
+          login,
+          spinner,
+          getEnvironmentVariables,
+        );
         const secret = options.envSecret || vars.FOREST_ENV_SECRET;
 
         if (!secret) {
@@ -64,14 +117,17 @@ export default function makeCommands({
         spinner.succeed('Environment found');
         spinner.stop();
 
-        const forestServer = buildHttpServer({ ...vars, FOREST_ENV_SECRET: secret });
+        const forestServer = validateAndBuildHttpServer(
+          { ...vars, FOREST_ENV_SECRET: secret },
+          buildHttpServer,
+        );
 
         if (!(await askToOverwriteCustomizations(spinner, forestServer))) {
           throw new BusinessError('Operation aborted');
         }
 
         spinner.start();
-        await bootstrap(secret, forestServer);
+        await bootstrap(secret, forestServer, bootstrapPathManager);
         spinner.succeed(
           'Project successfully bootstrapped. You can start creating your customizations!',
         );
@@ -82,11 +138,11 @@ export default function makeCommands({
     .command('login')
     .description('Login to your project')
     .action(
-      actionRunner(buildSpinner, async spinner => {
+      actionRunner(spinner, async () => {
         spinner.start('Logging in');
         const vars = await getEnvironmentVariables();
         validateServerUrl(vars.FOREST_SERVER_URL);
-        await login();
+        await login(spinner);
         spinner.succeed('You are now logged in');
       }),
     );
@@ -96,9 +152,13 @@ export default function makeCommands({
     .description('Publish your code customizations')
     .option('-f, --force', 'Force the publication without asking for confirmation')
     .action(
-      actionRunner(buildSpinner, async (spinner, options: { force: boolean }) => {
-        const vars = await getOrRefreshEnvironmentVariables();
-        const forestServer = buildHttpServer(vars);
+      actionRunner(spinner, async (options: { force: boolean }) => {
+        const vars = await getOrRefreshEnvironmentVariables(
+          login,
+          spinner,
+          getEnvironmentVariables,
+        );
+        const forestServer = validateAndBuildHttpServer(vars, buildHttpServer);
 
         if (!options.force && !(await askToOverwriteCustomizations(spinner, forestServer))) {
           throw new BusinessError('Operation aborted');
@@ -106,7 +166,7 @@ export default function makeCommands({
 
         spinner.start('Publishing code customizations (operation cannot be cancelled)');
         const subscriptionId = await publish(forestServer);
-        const subscriber = buildEventSubscriber(vars);
+        const subscriber = validateAndBuildEventSubscriber(vars, buildEventSubscriber);
 
         try {
           const { error } = await subscriber.subscribeToCodeCustomization(subscriptionId);
@@ -128,7 +188,7 @@ export default function makeCommands({
     .command('package')
     .description('Package your code customizations')
     .action(
-      actionRunner(buildSpinner, async spinner => {
+      actionRunner(spinner, async () => {
         spinner.start('Packaging code');
         await packageCustomizations();
         spinner.succeed('Code customizations packaged and ready for publish');