+++ packages/cloud-toolkit/src/services/environment-variables.ts
@@ -3,7 +3,6 @@ import { readFile } from 'node:fs/promises';
 import { homedir } from 'node:os';
 import path from 'path';
 
-import login from './login';
 import { BusinessError } from '../errors';
 import { EnvironmentVariables } from '../types';
 
@@ -79,14 +78,3 @@ export function validateEnvironmentVariables(env: EnvironmentVariables): void {
   validateServerUrl(env.FOREST_SERVER_URL);
   validateSubscriptionUrl(env.FOREST_SUBSCRIPTION_URL);
 }
-
-export const getOrRefreshEnvironmentVariables = async (): Promise<EnvironmentVariables> => {
-  let vars = await getEnvironmentVariables();
-
-  if (!vars.FOREST_AUTH_TOKEN) {
-    await login();
-    vars = await getEnvironmentVariables();
-  }
-
-  return vars;
-};