+++ packages/cloud-toolkit/test/commands/bootstrap.test.ts
@@ -4,14 +4,72 @@ import CommandTester from './command-tester';
 import { setupCommandArguments } from './utils';
 
 describe('bootstrap command', () => {
-  beforeEach(async () => {
-    await fs.rm('cloud-customizer', { force: true, recursive: true });
+  describe('when it is the first time to run bootstrap', () => {
+    it('should generate the cloud customizer folder', async () => {
+      const defaultEnvs = {
+        FOREST_SERVER_URL: 'https://api.forestadmin.com',
+        FOREST_SUBSCRIPTION_URL: 'wss://api.forestadmin.com/subscriptions',
+      };
+      const getEnvironmentVariables = jest
+        .fn()
+        .mockResolvedValueOnce({
+          FOREST_AUTH_TOKEN: null, // auth token is missing because login is required
+          ...defaultEnvs,
+        })
+        .mockResolvedValueOnce({
+          FOREST_AUTH_TOKEN: 'a-token-after-login', // auth token is present after login
+          ...defaultEnvs,
+        });
+
+      const getIntrospection = jest.fn().mockResolvedValue([
+        {
+          name: 'forestCollection',
+          schema: 'public',
+          columns: [
+            {
+              type: { type: 'scalar', subType: 'NUMBER' },
+              autoIncrement: true,
+              defaultValue: null,
+              isLiteralDefaultValue: true,
+              name: 'id',
+              allowNull: false,
+              primaryKey: true,
+              constraints: [],
+            },
+          ],
+          unique: [['id'], ['title']],
+        },
+      ]);
+
+      const setup = setupCommandArguments({ getIntrospection, getEnvironmentVariables });
+      const cloudCustomizerPath = setup.bootstrapPathManager.cloudCustomizer;
+      await fs.rm(cloudCustomizerPath, { force: true, recursive: true });
+
+      const cmd = new CommandTester(setup, [
+        'bootstrap',
+        '--env-secret',
+        'd4ae505b138c30f2d70952421d738627d65ca5322a27431d067479932cebcfa2',
+      ]);
+      await cmd.run();
+
+      expect(cmd.outputs).toEqual([
+        cmd.start('Bootstrapping project'),
+        cmd.succeed('Environment found'),
+        cmd.start('Bootstrapping project'),
+        cmd.succeed(
+          'Project successfully bootstrapped. You can start creating your customizations!',
+        ),
+      ]);
+
+      expect(setup.login).toHaveBeenCalled();
+      expect(await fs.stat(cloudCustomizerPath)).toBeTruthy();
+    });
   });
 
   describe('when forest env secret is missing', () => {
     it('should throw an error', async () => {
-      const getOrRefreshEnvironmentVariables = jest.fn().mockResolvedValue({});
-      const setup = setupCommandArguments({ getOrRefreshEnvironmentVariables });
+      const getEnvironmentVariables = jest.fn().mockResolvedValue({});
+      const setup = setupCommandArguments({ getEnvironmentVariables });
 
       const cmd = new CommandTester(setup, ['bootstrap']);
       await cmd.run();
@@ -28,10 +86,10 @@ describe('bootstrap command', () => {
 
   describe('when there is already a cloud customizer folder', () => {
     it('should throw an error', async () => {
-      const getOrRefreshEnvironmentVariables = jest.fn().mockResolvedValue({});
-      const setup = setupCommandArguments({ getOrRefreshEnvironmentVariables });
-
-      await fs.mkdir('cloud-customizer');
+      const setup = setupCommandArguments();
+      const cloudCustomizerPath = setup.bootstrapPathManager.cloudCustomizer;
+      await fs.rm(cloudCustomizerPath, { force: true, recursive: true });
+      await fs.mkdir(cloudCustomizerPath);
 
       const cmd = new CommandTester(setup, [
         'bootstrap',