+++ packages/cloud-toolkit/test/commands/login.test.ts
@@ -1,0 +1,73 @@
+import CommandTester from './command-tester';
+import { setupCommandArguments } from './utils';
+
+describe('login command', () => {
+  it('should login correctly', async () => {
+    const getEnvironmentVariables = jest
+      .fn()
+      .mockReturnValue({ FOREST_SERVER_URL: 'https://some.wrong.url' });
+    const login = jest.fn();
+    const setup = setupCommandArguments({ getEnvironmentVariables, login });
+
+    const cmd = new CommandTester(setup, ['login']);
+    await cmd.run();
+
+    expect(login).toHaveBeenCalled();
+    expect(cmd.outputs).toEqual([cmd.start('Logging in'), cmd.succeed('You are now logged in')]);
+  });
+
+  describe('when forest server url is missing', () => {
+    it('should throw an error', async () => {
+      const getEnvironmentVariables = jest.fn().mockReturnValue({});
+      const setup = setupCommandArguments({ getEnvironmentVariables });
+
+      const cmd = new CommandTester(setup, ['login']);
+      await cmd.run();
+
+      expect(cmd.outputs).toEqual([
+        cmd.start('Logging in'),
+        cmd.fail('Missing FOREST_SERVER_URL. Please check your .env file.'),
+      ]);
+    });
+  });
+
+  describe('when server URL is invalid', () => {
+    it('should throw an error', async () => {
+      const getEnvironmentVariables = jest
+        .fn()
+        .mockReturnValue({ FOREST_SERVER_URL: 'some.wrong.url' });
+
+      const setup = setupCommandArguments({ getEnvironmentVariables });
+
+      const cmd = new CommandTester(setup, ['login']);
+      await cmd.run();
+
+      expect(cmd.outputs).toEqual([
+        cmd.start('Logging in'),
+        cmd.fail(
+          'FOREST_SERVER_URL is invalid. Please check your .env file. ' +
+            'You can probably remove it from your .env file.: Invalid URL',
+        ),
+      ]);
+    });
+  });
+
+  describe("when server URL doesn't start with http", () => {
+    it('should throw an error', async () => {
+      const getEnvironmentVariables = jest
+        .fn()
+        .mockReturnValue({ FOREST_SERVER_URL: 'ws://some.wrong.url' });
+      const setup = setupCommandArguments({ getEnvironmentVariables });
+
+      const cmd = new CommandTester(setup, ['login']);
+      await cmd.run();
+
+      expect(cmd.outputs).toEqual([
+        cmd.start('Logging in'),
+        cmd.fail(
+          'FOREST_SERVER_URL is invalid, it must start with http:// or https://. Please check your .env file.',
+        ),
+      ]);
+    });
+  });
+});